lecture4
Сделал отдельно 'user' и 'auth'
