+++ src/store/exports.js
@@ -4,4 +4,5 @@ export { default as modals } from './modals';
 export { default as article } from './article';
 export { default as locale } from './locale';
 export { default as user } from './user';
+export { default as auth } from './auth';
 export { default as categories } from './categories';
